+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: build clean deploy
 
 build:
-	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o bin/main cmd/*.go
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o bin/api/main ./cmd/api/...
 
 docs: build
 	./bin/main -gendocs
